requirements.txt
dgl==2.4.0+cu124
ImageHash==4.3.2
imageio==2.37.2
lightning==2.5.2
matplotlib==3.10.7
numpy==2.3.5
opencv_contrib_python==4.12.0.88
pandas==2.3.3
Pillow==12.0.0
PyYAML==6.0.2
PyYAML==6.0.3
scikit_learn==1.7.2
scipy==1.16.3
skimage==0.0
torch==2.9.0
torch==2.6.0
torch==2.4.0
torchmetrics==1.2.1
torchvision==0.24.0
torchvision==0.15.2a0
tqdm==4.67.1